+++ 15-consumer-configuration/benthos/seed.yaml
@@ -1,0 +1,25 @@
+input:
+  generate:
+    count: 10000
+    interval: ""
+    ## Subject template "orders.{location}.{id}.{status}"
+    mapping: |
+      let locations = ["US", "CA", "EU"]
+      root.id = random_int()
+      root.location = $locations.index(random_int(max: 2))
+      root.status = "pending"
+      root.created_at = now()
+      root.sku = uuid_v4()
+      root.customer = {}
+      root.customer.first_name = fake("first_name")
+      root.customer.last_name = fake("first_name")
+      root.customer.address = fake("phone_number")
+      # root = root.format_json()
+
+output:
+  label: "event_publisher"
+  nats_jetstream:
+    urls: ["nats://connect.ngs.global"]
+    subject: orders.${!location}.${!id}.${!status}
+    auth:
+      user_credentials_file: user.creds
