+++ 15-consumer-configuration/benthos/shipped.yaml
@@ -1,0 +1,27 @@
+input:
+  label: orders
+  nats_jetstream:
+    stream: orders
+    subject: orders.*.*.confirmed
+    urls: ["nats://connect.ngs.global"]
+    auth:
+      user_credentials_file: user.creds
+
+pipeline:
+  processors:
+    - mapping: |
+        root = this
+        root.status = if random_int() % 2 == 0 {
+          "shipped"
+        } else {
+          "cancelled"
+        }
+        root.created_at = now()
+
+output:
+  label: "event_publisher"
+  nats_jetstream:
+    urls: ["nats://connect.ngs.global"]
+    subject: orders.${!location}.${!id}.${!status}
+    auth:
+      user_credentials_file: user.creds
